Added cast to int for hermit.hermit_values.n_cols in operator<<

+++ src/Hermit.cpp
@@ -39,8 +39,8 @@ arma::colvec Hermit::get(int n) {
 }
 
 std::ostream& operator<<(std::ostream &stream, const Hermit &hermit) {
-    for (int z = 0; z < hermit.hermit_values.n_cols; ++z) {
-        for (int n = 0; n < hermit.hermit_values.n_rows; ++n) {
+    for (int z = 0; z < (int) hermit.hermit_values.n_cols; ++z) {
+        for (int n = 0; n < (int) hermit.hermit_values.n_rows; ++n) {
             stream /*<< std::setfill('.')*/ << std::setw(hermit.insert_size) << hermit.hermit_values(n, z) << " ";
         }
         stream << std::endl;
